+++ src/index.js
@@ -3,7 +3,7 @@ import renderSelectors from './core/SelectorDropdowns';
 // import renderEChart from './core/charts/ExampleChart';
 // import renderTable from './core/tables/ExampleTable';
 import renderEducationCharts from './core/charts/EducationCharts';
-import renderTable from './core/tables/NumberOfSchoolsTable';
+import renderEducationTable from './core/tables/EducationTables';
 import './styles/styles.css';
 import renderKeyFacts from './core/KeyFacts';
 
@@ -13,7 +13,7 @@ import renderKeyFacts from './core/KeyFacts';
 window.addEventListener('load', () => {
   renderSelectors('district-selectors');
   // renderEChart();
-  renderTable('dicharts--table--number-of-schools');
+  renderEducationTable('dicharts--table--number-of-schools');
   renderKeyFacts('dicharts--keyfacts');
   renderEducationCharts();
 });
